f138138 Daemon: fix hashrate display in diff command output (iDunk5400)

+++ src/daemon/rpc_command_executor.cpp
@@ -261,7 +261,7 @@ bool t_rpc_command_executor::show_difficulty() {
   tools::success_msg_writer() <<   "BH: " << res.height
                               << ", TH: " << res.top_block_hash
                               << ", DIFF: " << res.difficulty
-                              << ", HR: " << (int) res.difficulty / res.target << " H/s";
+                              << ", HR: " << res.difficulty / res.target << " H/s";
 
   return true;
 }
